+++ regression/run/funptr/mainc.c
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2017-2024, Lawrence Livermore National Security, LLC and
+ * other Shroud Project Developers.
+ * See the top-level COPYRIGHT file for details.
+ *
+ * SPDX-License-Identifier: (BSD-3-Clause)
+ * #######################################################################
+ *
+ * Test C API generated from clibrary.yaml.
+ */
+
+#include "clibrary.h"
+
+
+int main(int argc, char *argv[])
+{
+
+}